Serve static files on root route
Add a route on the root (`/`) of the Express app to serve
static files. The idea is being able to serve both the frontend
static files and the backend APIs through the same server.

By default, the static files will be served from the `build` dir
within the `labephoto-front-end` repo. The code assumes that the
frontend repo is cloned beside the backend repo. The static dir
to be served can be overriden with the `STATIC_DIR` env var.